2016031101 标准适配修改

+++ commons-rest-model/src/main/java/com/yihu/ehr/model/adaption/MOrgDictItem.java
@@ -20,7 +20,7 @@ public class MOrgDictItem {
     String description;
     String createUser;
     String updateUser;
-    Integer orgDict;
+    long orgDict;
     String organization;
     int sequence;
 
@@ -96,11 +96,11 @@ public class MOrgDictItem {
         this.updateUser = updateUser;
     }
 
-    public Integer getOrgDict() {
+    public long getOrgDict() {
         return orgDict;
     }
 
-    public void setOrgDict(Integer orgDict) {
+    public void setOrgDict(long orgDict) {
         this.orgDict = orgDict;
     }
